Move app-owned components and helpers out of registry-managed paths
confirm-dialog, split-button and user-badge are nebi-specific wrappers, not
@nebari registry items, so they now live in src/components/ instead of
src/components/ui/. Likewise capitalize and the status colour helpers move
from src/lib/utils.ts (registry-owned, cn() only) into src/lib/strings.ts and
src/lib/status.ts. Importers and tests follow; split-button gains the explicit
button type Biome requires now that it is linted.
