+++ frontend/src/lib/strings.ts
@@ -1,0 +1,4 @@
+export function capitalize(str: string): string {
+  if (!str) return str;
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
