+++ frontend/src/lib/utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { capitalize, cn } from './utils';
+import { cn } from './utils';
 
 describe('cn', () => {
   it('merges class names', () => {
@@ -23,27 +23,5 @@ describe('cn', () => {
 
   it('returns empty string when called with no arguments', () => {
     expect(cn()).toBe('');
-  });
-});
-
-describe('capitalize', () => {
-  it('capitalizes the first letter', () => {
-    expect(capitalize('hello')).toBe('Hello');
-  });
-
-  it('leaves an already-capitalized string unchanged', () => {
-    expect(capitalize('Hello')).toBe('Hello');
-  });
-
-  it('handles a single character', () => {
-    expect(capitalize('a')).toBe('A');
-  });
-
-  it('returns empty string unchanged', () => {
-    expect(capitalize('')).toBe('');
-  });
-
-  it('does not alter the rest of the string', () => {
-    expect(capitalize('hELLO WORLD')).toBe('HELLO WORLD');
   });
 });
